requirements.txt
flask==3.1.2
tensorflow==2.20.0
numpy>=2.0.0,<2.3.0
opencv-python==4.12.0.88
pillow==12.0.0
pandas==2.3.3
scikit-learn==1.7.2

